PR for Node API version? #101
