+++ package.json
@@ -1,54 +1,55 @@
 {
-  "name": "concurrently",
-  "version": "3.5.0",
-  "description": "Run commands concurrently",
-  "main": "src/main.js",
-  "bin": {
-    "concurrent": "./src/main.js",
-    "concurrently": "./src/main.js"
-  },
-  "engines": {
-    "node": ">=4.0.0"
-  },
-  "scripts": {
+    "name": "concurrently",
+    "version": "3.5.0",
+    "description": "Run commands concurrently",
+    "main": "src/api.js",
+    "typings": "typings.d.ts",
+    "bin": {
+        "concurrent": "./src/main.js",
+        "concurrently": "./src/main.js"
+    },
+    "engines": {
+        "node": ">=4.0.0"
+    },
+    "scripts": {
     "test": "mocha"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/kimmobrunfeldt/concurrently.git"
-  },
-  "keywords": [
-    "bash",
-    "concurrent",
-    "parallel",
-    "concurrently",
-    "command",
-    "sh"
-  ],
-  "author": "Kimmo Brunfeldt",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/kimmobrunfeldt/concurrently/issues"
-  },
-  "homepage": "https://github.com/kimmobrunfeldt/concurrently",
-  "dependencies": {
-    "chalk": "0.5.1",
-    "commander": "2.6.0",
-    "date-fns": "^1.23.0",
-    "lodash": "^4.5.1",
-    "rx": "2.3.24",
-    "spawn-command": "^0.0.2-1",
-    "supports-color": "^3.2.3",
-    "tree-kill": "^1.1.0"
-  },
-  "devDependencies": {
-    "chai": "^1.10.0",
-    "mocha": "^2.1.0",
-    "mustache": "^1.0.0",
-    "releasor": "^1.2.1",
-    "semver": "^4.2.0",
-    "shell-quote": "^1.4.3",
-    "shelljs": "^0.3.0",
-    "string": "^3.0.0"
-  }
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/kimmobrunfeldt/concurrently.git"
+    },
+    "keywords": [
+        "bash",
+        "concurrent",
+        "parallel",
+        "concurrently",
+        "command",
+        "sh"
+    ],
+    "author": "Kimmo Brunfeldt",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/kimmobrunfeldt/concurrently/issues"
+    },
+    "homepage": "https://github.com/kimmobrunfeldt/concurrently",
+    "dependencies": {
+        "chalk": "0.5.1",
+        "commander": "2.6.0",
+        "date-fns": "^1.23.0",
+        "lodash": "^4.5.1",
+        "rx": "2.3.24",
+        "spawn-command": "^0.0.2-1",
+        "supports-color": "^3.2.3",
+        "tree-kill": "^1.1.0"
+    },
+    "devDependencies": {
+        "chai": "^1.10.0",
+        "mocha": "^2.1.0",
+        "mustache": "^1.0.0",
+        "releasor": "^1.2.1",
+        "semver": "^4.2.0",
+        "shell-quote": "^1.4.3",
+        "shelljs": "^0.3.0",
+        "string": "^3.0.0"
+    }
 }
